Update: changed the speed for every LED stage, going to transfer it onto the robotics computer to have it tested
