+++ src/main/java/frc/robot/Constants.java
@@ -18,13 +18,13 @@ import edu.wpi.first.math.util.Units;
 import java.util.Map;
 
 public final class Constants {
-  public static final boolean kDebugEnabled = true;
+  public static final boolean kDebugEnabled = false;
 
-  public static final boolean kIntakeEnabled = true;
-  public static final boolean kElevatorEnabled = true;
-  public static final boolean kArmEnabled = true;
-  public static final boolean kSwerveEnabled = true;
-  public static final boolean kLedStripEnabled = false;
+  public static final boolean kIntakeEnabled = false;
+  public static final boolean kElevatorEnabled = false;
+  public static final boolean kArmEnabled = false;
+  public static final boolean kSwerveEnabled = false;
+  public static final boolean kLedStripEnabled = true;
 
   public static final RobotType kRobotType = RobotType.ALPHA;
   public static final RobotMode kCurrentMode = RobotMode.SIM;
