+++ src/main/java/frc/robot/led/commands/LEDSetAllSectionsPattern.java
@@ -30,4 +30,9 @@ public class LEDSetAllSectionsPattern extends CommandBase {
   public boolean isFinished() {
     return true;
   }
+
+  @Override
+  public boolean runsWhenDisabled() {
+    return true;
+  }
 }
