+++ src/main/java/frc/robot/led/patterns/LockSwervePattern.java
@@ -16,6 +16,6 @@ import frc.robot.led.patternBases.LEDPattern;
 public class LockSwervePattern extends LEDPattern {
   public LockSwervePattern() {
     super();
-    setPixelRange(1, kResolution, Color.kDarkRed);
+    setPixelRange(1, kResolution, Color.kRed);
   }
 }
